Address last PR comments & add changeset
