+++ .changeset/selfish-terms-agree.md
@@ -1,0 +1,5 @@
+---
+'@mysten/sui': minor
+---
+
+Adds experimental named packages plugin
