+++ sdk/typescript/src/transactions/index.ts
@@ -52,3 +52,5 @@ export {
 	namedPackagesPlugin,
 	type NamedPackagesPluginOptions,
 } from './plugins/NamedPackagesPlugin.js';
+
+export { type NamedPackagesPluginCache } from './plugins/utils.js';
